Fix build error with snapshot version bump
Issue-ID: NFVO-126
After bumping the version from 1.0.1-SNAPSHOT to 1.1.0-SNAPSHOT,
there will be a build error caused by the version bump, the
binary of old version is no longer there in the Nexus server.
This fix changes the version of dependency package to new
version and fix the build error.

Also fixed the coding format issue to change [Tab] with space
and remove the ending [Tab] or space.

Change-Id: I598ccbed9a0e33d2ef7cc91aa906af60796abac0
Signed-off-by: Johnson Li <johnson.li@intel.com>